# openapi.yaml
openapi: 3.0.3
info:
  title: Face Verification Backbone API
  version: "1.0.0"
  description: >
    Identity verification as a service. Two integration models:
    **managed** (we store encrypted templates per tenant) and **stateless**
    (you keep your own data and pass it in). All endpoints except /v1/health
    require an API key sent as the `X-API-Key` header. Verification/compare
    results are HMAC-signed with your tenant's signing secret so your app can
    trust the allow/deny outcome.


    Keys carry a **role**: `admin` (full control) or `verify` (recognition only —
    cannot enrol, delete, or list). Each tenant is isolated. Calls are metered
    per tenant (see `/v1/usage`) and may be subject to a monthly quota
    (HTTP 429 `quota_exceeded`).
servers:
  - url: https://your-host:5000
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Signature:
      type: object
      properties:
        alg: { type: string, example: HMAC-SHA256 }
        ts: { type: string }
        nonce: { type: string }
        hmac: { type: string }
security:
  - ApiKeyAuth: []
paths:
...
  /v1/enroll:
    post:
      summary: "Managed: enroll a user from one or more images"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string }
                image: { type: string }
                images: { type: array, items: { type: string } }
      responses: { "200": { description: Enrolment summary } }
...
  /v1/verify:
    post:
      summary: "Managed: verify a claimed user (1:1) or identify (1:N)"
      description: >
        Pass `image` for a single-shot check, or `frames` + `token` (from
        /v1/challenge) for an active-liveness head-turn check. Omit `user_id`
        to identify (1:N).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                image: { type: string }
                frames: { type: array, items: { type: string } }
                token: { type: string }
      responses:
        "200":
          description: Verification result (signed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, description: granted }
                  user_id: { type: string }
                  score: { type: number }
                  signature: { $ref: "#/components/schemas/Signature" }
  /v1/identify:
    post:
      summary: "Managed: 1:N identify (image or frames+token)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image: { type: string }
                frames: { type: array, items: { type: string } }
                token: { type: string }
      responses: { "200": { description: Identify result (signed) } }
...
  /v1/users/delete:
    post:
      summary: "Managed: delete one user or many (role: delete)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                user_ids: { type: array, items: { type: string } }
      responses: { "200": { description: Deletion result } }
...